+++ src/object/object.tests.ts
@@ -1,5 +1,8 @@
 import o from 'ospec';
-import { objectClean } from './object';
+import beget from './beget';
+import { objectClean, objectIsSame } from './object';
+
+// ===========================================================================
 
 o.spec('objectClean', () => {
 	type T1 = {
@@ -47,6 +50,8 @@ o.spec('objectClean', () => {
 	> = true;
 	/* eslint-enable @typescript-eslint/no-unused-vars */
 
+	// -------------------------------------
+
 	o('Removes undefineds', () => {
 		const obj = { a: 1, b: undefined, c: null };
 		const cleaned = objectClean(obj);
@@ -78,5 +83,58 @@ o.spec('objectClean', () => {
 		const cleaned = objectClean(obj);
 		o(cleaned).equals(obj);
 		o(Object.keys(cleaned).length).equals(2);
+	});
+});
+
+// ===========================================================================
+
+o.spec('objectIsSame', () => {
+	o('simple case', () => {
+		const xy = { x: 1, y: 2 };
+		const xy2 = { x: 1, y: 2 };
+		const xyz = { x: 1, y: 2, z: 3 };
+
+		o(objectIsSame(xy, xy)).equals(true)('literally same');
+		o(objectIsSame(xy, xy2)).equals(true)('same content');
+		o(objectIsSame(xy, xyz)).equals(false)('B has extra properties');
+		o(objectIsSame(xyz, xy)).equals(false)('B is missing properties');
+	});
+
+	o('works for Arrays', () => {
+		const arr = [5, undefined, 2];
+		o(objectIsSame(arr, arr)).equals(true)('literally same');
+		o(objectIsSame(arr, [5, undefined, 2])).equals(true)('same values');
+		o(objectIsSame(arr, [undefined, 5, 2])).equals(false)('same values, different order');
+		o(objectIsSame([99], [99, undefined])).equals(false)('different length');
+
+		const arrCloned = [].concat.call([], arr);
+		arrCloned.length = 10;
+		o(objectIsSame(arr, arrCloned)).equals(true)(
+			'Array#length does not throw off the comparison'
+		);
+	});
+
+	o('captures explicly undefined values', () => {
+		o(objectIsSame({ x: undefined }, { y: 2 })).equals(false)('literally same');
+	});
+
+	o('ignores prototype chain properties', () => {
+		const xy = { x: 1, y: 2 };
+		const xyz = { x: 1, y: 2, z: 3 };
+		const xyInherited = beget(xy);
+		const xy_zInherited = beget(xyz, { x: 1, y: 2 });
+
+		o(objectIsSame(xyInherited, xy)).equals(false)('A inherits properties ');
+		o(objectIsSame(xy, xyInherited)).equals(false)('B inherits properties');
+		o(objectIsSame(xy, xy_zInherited)).equals(true)('B has extra inherited properties');
+		o(objectIsSame(xy_zInherited, xy)).equals(true)('B is missing inherited properties');
+	});
+
+	o('works for Arrays', () => {
+		const arr = [5, undefined, 2];
+		o(objectIsSame(arr, arr)).equals(true)('literally same');
+		o(objectIsSame(arr, [5, undefined, 2])).equals(true)('same values');
+		o(objectIsSame(arr, [undefined, 5, 2])).equals(false)('same values, different order');
+		o(objectIsSame([99], [99, undefined])).equals(false)('different length');
 	});
 });